remove duplicate progress.service
